feat: add sbt build
* add sbt build
* split analyzer into small files
* several import optimize
