+++ src/main/scala/io/github/jbellis/brokk/IAnalyzer.scala
@@ -1,7 +1,6 @@
 package io.github.jbellis.brokk
 
 import java.util
-import java.util.List
 
 trait IAnalyzer {
   def getAllClasses: util.List[CodeUnit]
