+++ src/main/scala/io/github/jbellis/brokk/Language.scala
@@ -1,0 +1,9 @@
+package io.github.jbellis.brokk
+
+sealed trait Language
+
+object Language {
+  case object Java extends Language
+
+  case object Python extends Language
+}
